Update README.md
LATM is a research, development and innovation laboratory for audio, sound computing and music technologies, linked to the School of Arts, Sciences and Humanities (EACH) at the University of São Paulo (USP).

In addition to research, the laboratory carries out extension activities, such as seminars and scientific events, and hosts educational activities, such as undergraduate and postgraduate courses, training and qualification, as well as activities aimed at sound creation, cultural and artistic production mediated by technologies. 

We defend ethical and humanistic values, and we seek to adopt altruistic attitudes and positions that defend well-being and evolution as guiding guidelines in the design of new technologies. This repository is a point of information about projects and distribution of developments carried out by the laboratory team and collaborators.